+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ### unreleased changes
 
-none
+
+Breaking changes:
+
+- grpcweb-transport deadline option is used as a timeout, but should be used as a timestamp (see #80)
+
 
 
 ### v2.0.0-alpha.12