+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ### unreleased changes
 
 
+none
+
+
+### v2.0.0-alpha.13
+
 Breaking changes:
 
 - grpcweb-transport deadline option is used as a timeout, but should be used as a timestamp (see #80)